Misc refactoring and bug fixes
