+++ lib/get_type.js
@@ -1,15 +1,17 @@
 'use strict';
 
 module.exports = function getType(val) {
-    if (val instanceof Number)
+    if (val instanceof Number) {
         return 'number';
-    if (val instanceof String)
+    } else if (val instanceof String) {
         return 'string';
-    if (val instanceof Boolean)
+    } else if (val instanceof Boolean) {
         return 'boolean';
-    if (Array.isArray(val))
+    } else if (Array.isArray(val)) {
         return 'array';
-    if (val === null)
+    } else if (val === null) {
         return 'null';
-    return typeof val;
+    } else {
+        return typeof val;
+    }
 };
