+++ lib/validate/validate.js
@@ -55,7 +55,9 @@ module.exports = function validate(options) {
             valueSpec: valueSpec,
             key: key,
             style: style,
-            styleSpec: styleSpec
+            styleSpec: styleSpec,
+            array: options.array,
+            arrayIndex: options.arrayIndex
         });
 
     } else if (valueSpec.type && VALIDATORS[valueSpec.type]) {
@@ -64,7 +66,9 @@ module.exports = function validate(options) {
             valueSpec: valueSpec,
             key: key,
             style: style,
-            styleSpec: styleSpec
+            styleSpec: styleSpec,
+            array: options.array,
+            arrayIndex: options.arrayIndex
         });
 
     } else {
@@ -73,7 +77,9 @@ module.exports = function validate(options) {
             valueSpec: valueSpec.type ? styleSpec[valueSpec.type] : valueSpec,
             key: key,
             style: style,
-            styleSpec: styleSpec
+            styleSpec: styleSpec,
+            array: options.array,
+            arrayIndex: options.arrayIndex
         });
     }
 };
