+++ lib/validate/validate_layer.js
@@ -23,14 +23,10 @@ module.exports = function validateLayer(options) {
     var ref = unbundle(layer.ref);
 
     if (layer.id) {
-        var isFirstMatch = true;
         for (var i = 0; i < options.arrayIndex; i++) {
             var otherLayer = style.layers[i];
             if (unbundle(otherLayer.id) === unbundle(layer.id)) {
-                if (!isFirstMatch) {
-                    errors.push(new ValidationError(key, layer.id, 'duplicate layer id "%s", used at line %d and line %d', layer.id, layer.id.__line__, otherLayer.id.__line__));
-                }
-                isFirstMatch = false;
+                errors.push(new ValidationError(key, layer.id, 'duplicate layer id "%s", previously used at line %d', layer.id, otherLayer.id.__line__));
             }
         }
     }
