+++ lib/validate_style.js
@@ -1,20 +1,17 @@
 'use strict';
 
-var validateConstants = require('./validate_constants');
-var validate = require('./validate');
-var latestStyleSpec = require('../../reference/latest.min');
-var parseStyle = require('../parse_style');
-var ParsingError = require('../parsing_error');
+var validateConstants = require('./validate/validate_constants');
+var validate = require('./validate/validate');
+var latestStyleSpec = require('../reference/latest.min');
+var parseStyle = require('./parse_style');
+var ParsingError = require('./parsing_error');
 
 module.exports = function(style, styleSpec) {
     try {
         return validateStyleObject(parseStyle(style), styleSpec);
     } catch (e) {
-        if (e instanceof ParsingError) {
-            return [e];
-        } else {
-            throw e;
-        }
+        if (e instanceof ParsingError) return [e];
+        else throw e;
     }
 };
