+++ test/fixture/layers.output.json
@@ -52,6 +52,10 @@
     "line": 64
   },
   {
+    "message": "layers[11]: duplicate layer id \"duplicate\", previously used at line 68",
+    "line": 74
+  },
+  {
     "message": "layers[12].type: expected one of [fill, line, symbol, circle, raster, background], invalid found",
     "line": 81
   },
