+++ test/validate.js
@@ -12,7 +12,8 @@ var UPDATE = !!process.env.UPDATE;
 glob.sync(__dirname + '/fixture/*.input.json').forEach(function(file) {
     t(path.basename(file), function(t) {
         var outputfile = file.replace('.input', '.output');
-        var result = validate(fs.readFileSync(file));
+        var style = fs.readFileSync(file);
+        var result = validate(style);
         if (UPDATE) fs.writeFileSync(outputfile, JSON.stringify(result, null, 2));
         var expect = JSON.parse(fs.readFileSync(outputfile));
         t.deepEqual(result, expect);
